+++ include/handler/handlers/connect.h
@@ -1,0 +1,9 @@
+#ifndef HANDLER_CONNECT_H
+#define HANDLER_CONNECT_H
+
+#include "request.h"
+#include "response.h"
+
+Response* connect(Request*);
+
+#endif
