+++ include/handler/handlers/dashboard.h
@@ -1,0 +1,9 @@
+#ifndef DASHBOARD_H
+#define DASHBOARD_H
+
+#include "request.h"
+#include "response.h"
+
+Response* dashboard(Request*);
+
+#endif
