+++ include/handler/handlers/home.h
@@ -1,0 +1,9 @@
+#ifndef HOME_H
+#define HOME_H
+
+#include "request.h"
+#include "response.h"
+
+Response* home(Request* req);
+
+#endif
