+++ include/handler/handlers/like.h
@@ -1,0 +1,9 @@
+#ifndef HANDLERS_LIKE_H
+#define HANDLERS_LIKE_H
+
+#include "request.h"
+#include "response.h"
+
+Response* like(Request* req);
+
+#endif
