+++ include/handler/handlers/login.h
@@ -1,0 +1,9 @@
+#ifndef HANDLERS_LOGIN_H
+#define HANDLERS_LOGIN_H
+
+#include "request.h"
+#include "response.h"
+
+Response* login(Request*);
+
+#endif
