+++ include/handler/handlers/logout.h
@@ -1,0 +1,9 @@
+#ifndef HANDLERS_LOGOUT_H
+#define HANDLERS_LOGOUT_H
+
+#include "request.h"
+#include "response.h"
+
+Response* logout(Request*);
+
+#endif
