+++ include/handler/handlers/not_found.h
@@ -1,0 +1,9 @@
+#ifndef NOT_FOUND_H
+#define NOT_FOUND_H
+
+#include "request.h"
+#include "response.h"
+
+Response* not_found(Request*);
+
+#endif
