+++ include/handler/handlers/post.h
@@ -1,0 +1,9 @@
+#ifndef HANDLER_POST_H
+#define HANDLER_POST_H
+
+#include "request.h"
+#include "response.h"
+
+Response* post(Request* req);
+
+#endif
