+++ include/handler/handlers/profile.h
@@ -1,0 +1,9 @@
+#ifndef PROFILE_H
+#define PROFILE_H
+
+#include "request.h"
+#include "response.h"
+
+Response* profile(Request* req);
+
+#endif
