+++ include/handler/handlers/search.h
@@ -1,0 +1,9 @@
+#ifndef FACEBOOC_SEARCH_H
+#define FACEBOOC_SEARCH_H
+
+#include "request.h"
+#include "response.h"
+
+Response* search(Request*);
+
+#endif	// FACEBOOC_SEARCH_H
