+++ include/handler/handlers/signup.h
@@ -1,0 +1,9 @@
+#ifndef HANDLERS_SIGNUP_H
+#define HANDLERS_SIGNUP_H
+
+#include "request.h"
+#include "response.h"
+
+Response* signup(Request*);
+
+#endif
