+++ include/handler/handlers/unlike.h
@@ -1,0 +1,9 @@
+#ifndef HANDLERS_UNLIKE_H
+#define HANDLERS_UNLIKE_H
+
+#include "request.h"
+#include "response.h"
+
+Response* unlike(Request*);
+
+#endif
